Fix spacing, move days_ago to utils
