+++ tle/cogs/codeforces.py
@@ -154,18 +154,10 @@ class Codeforces(commands.Cog):
         else:
             submissions.sort(key=lambda sub: sub.creationTimeSeconds, reverse=True)
 
-        def days_ago(t):
-            days = (time.time() - t)/(60*60*24)
-            if days <= 1:
-                return 'today'
-            if days <= 2:
-                return 'yesterday'
-            return f'{math.ceil(days)} days ago'
-
         msg = '\n'.join(
-                f'[{sub.problem.name}]({sub.problem.url})' +
-                f'[{sub.problem.rating if sub.problem.rating else "?"}]' +
-                f'({days_ago(sub.creationTimeSeconds)})'
+                f'[{sub.problem.name}]({sub.problem.url}) '
+                f'[{sub.problem.rating if sub.problem.rating else "?"}] '
+                f'({cf_common.days_ago(sub.creationTimeSeconds)})'
                 for sub in submissions[:10]
         )
         title = '{} solved {} problems by `{}`'.format('Hardest' if hardest else 'Recently',
